fix dsco error

+++ modules/consensus/maintenance.go
@@ -44,7 +44,26 @@ func applyMaturedSiacoinOutputs(tx *bolt.Tx, pb *processedBlock) {
 		return
 	}
 
-	_ = forEachDSCO(tx, pb.Height, func(id types.SiacoinOutputID, sco types.SiacoinOutput) error {
+	// Iterate through the list of delayed siacoin outputs. Sometimes boltdb
+	// has trouble if you delete elements in a bucket while iterating through
+	// the bucket (and sometimes not - nondeterministic), so all of the
+	// elements are collected into an array and then deleted after the bucket
+	// scan is complete.
+	bucketID := append(prefixDSCO, encoding.Marshal(pb.Height)...)
+	var dscods []modules.DelayedSiacoinOutputDiff
+	dbErr := tx.Bucket(bucketID).ForEach(func(idBytes, scoBytes []byte) error {
+		// Decode the key-value pair into an id and a siacoin output.
+		var id types.SiacoinOutputID
+		var sco types.SiacoinOutput
+		encErr := encoding.Unmarshal(idBytes, &id)
+		if build.DEBUG && encErr != nil {
+			panic(encErr)
+		}
+		encErr = encoding.Unmarshal(scoBytes, &sco)
+		if build.DEBUG && encErr != nil {
+			panic(encErr)
+		}
+
 		// Sanity check - the output should not already be in siacoinOuptuts.
 		if build.DEBUG && isSiacoinOutput(tx, id) {
 			panic(errOutputAlreadyMature)
@@ -60,17 +79,24 @@ func applyMaturedSiacoinOutputs(tx *bolt.Tx, pb *processedBlock) {
 		pb.SiacoinOutputDiffs = append(pb.SiacoinOutputDiffs, scod)
 		commitSiacoinOutputDiff(tx, scod, modules.DiffApply)
 
-		// Remove the delayed siacoin output from the consensus set.
+		// Create the dscod and add it to the list of dscods that should be
+		// deleted.
 		dscod := modules.DelayedSiacoinOutputDiff{
 			Direction:      modules.DiffRevert,
 			ID:             id,
 			SiacoinOutput:  sco,
 			MaturityHeight: pb.Height,
 		}
-		pb.DelayedSiacoinOutputDiffs = append(pb.DelayedSiacoinOutputDiffs, dscod)
-		commitDelayedSiacoinOutputDiff(tx, dscod, modules.DiffApply)
+		dscods = append(dscods, dscod)
 		return nil
 	})
+	if build.DEBUG && dbErr != nil {
+		panic(dbErr)
+	}
+	for _, dscod := range dscods {
+		pb.DelayedSiacoinOutputDiffs = append(pb.DelayedSiacoinOutputDiffs, dscod)
+		commitDelayedSiacoinOutputDiff(tx, dscod, modules.DiffApply)
+	}
 	deleteDSCOBucket(tx, pb.Height)
 }
